+++ src/compatibility/eventListeners.ts
@@ -1,11 +1,10 @@
-import { Notice, TFile } from 'obsidian';
+import { TFile } from 'obsidian';
 import BetterDailyNotePlugin from '../main';
 import { createNotice } from '../utils';
 import { getSettingsFromExternalPlugin, moveDailyNote } from './utils';
 
 async function checkExternalPlugins(plugin: BetterDailyNotePlugin): Promise<string[]> {
     if (plugin.settings.compatibleDateFormats.length === 0) return [];
-    console.log(plugin.settings.compatibleDateFormats);
     if (plugin.settings.compatibleDateFormats.includes("AUTO")) {
         const externalFormats = await getSettingsFromExternalPlugin(plugin.app, plugin.settings);
         plugin.settings.compatibleDateFormats = Array.from(externalFormats);